+++ test/formatters/installers/buildifier.bash
@@ -1,0 +1,2 @@
+version=6.0.1
+curl -L --output /usr/local/bin/buildifier "https://github.com/bazelbuild/buildtools/releases/download/$version/buildifier-linux-amd64" && chmod +x /usr/local/bin/buildifier